requirements.txt
django
djangorestframework
environs
django-filter

venv=api1venv
username=APITutorial
password=tutorialpassword